Video Upload
The frontend component VideoUpload.js allows users to select a video file and triggers the handleVideoUpload function when the form is submitted. The video file is then sent to the backend using a POST request to the /api/upload-video endpoint.


On the backend, the /api/upload-video endpoint checks if a video file is present in the request and saves it to a local directory (in this case, the ./videos/ directory). You would